Move out first timeline submodule all timeline ref/guard entities
