+++ pageserver/src/tenant.rs
@@ -91,8 +91,8 @@ mod timeline;
 pub mod size;
 
 pub use timeline::{
-    with_ondemand_download, PageReconstructError, PageReconstructResult, Timeline, TimelineGuard,
-    TimelineRef,
+    timeline_ref::{TimelineGuard, TimelineRef},
+    with_ondemand_download, PageReconstructError, PageReconstructResult, Timeline,
 };
 
 // re-export this function so that page_cache.rs can use it.
